pyproject.toml
[project]
name = "sector-etf"
version = "0.1.0"
description = "Sector ETF daily snapshot fetcher"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests==2.32.3",
    "beautifulsoup4==4.12.3",
    "supabase==2.6.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = []
